+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/Drivetrain/IDrivetrain.java
@@ -10,45 +10,39 @@ public interface IDrivetrain {
 
     /**
      *
-     * @param highPower The maximum speed the robot will move at
-     * @param lowPower The lowest speed the robot will move at
-     * @param powerChange The distance or time the robot is away from the target
-     * @param powerGain How much the power change affects the speed (speed = powerChange*powerGain but has to be inbetween low and high power)
-     * @param moveAngle The angle at which the robot will move in the frame of reference of the driver
-     * @param oGain How much
-     * @param pGain
-     * @param endOrientationAngle
-     * @param endCondition
-     * @param timeAfterAngle
+     * @param currentPosition The current position of robot in any unit eg. encoder counts, sensor distances...
+     * @param targetPosition The target position of the robot in the same unit as current position
+     * @param rampDownTargetPosition Position at which the robot will start ramping down
+     * @param rampUpTargetPosition Position at which the robot will stop ramping up in power
+     * @param maxPower The maximum power the robot will move at
+     * @param lowPower The lowest power the robot will move at
+     * @param moveAngle The angle at which the robot will move in the frame of reference of the starting position
+     * @param PIDGain Three gains to control PID feedback loop for Orientation correction
+     * @param endOrientationAngle The direction the robot is facing
+     * @param timeAfterAngle How Much time the robot spends correcting after a pivot
      * @return
      */
-    boolean moveIMU(double highPower, double lowPower, double powerChange, double powerGain, double moveAngle, double oGain, double pGain, double endOrientationAngle, boolean endCondition, double timeAfterAngle);
+    boolean moveIMU(double currentPosition, double targetPosition, double rampDownTargetPosition, double rampUpTargetPosition, double maxPower, double lowPower, double moveAngle, double[] PIDGain, double endOrientationAngle, double timeAfterAngle);
 
 
 
-    boolean moveNoIMU(double angle, double speed, boolean condition, double pivotAmount);
-    /**
-     * balances the robot on the balancing stone
-     * @param highPower The maximum speed the robot will move at
-     * @param lowPower The lowest speed the robot will move at
-     * @param moveAngle the angle at which the robot maintain on the balancing board
-     * @param oGain gain to maintain angle
-     * @param endCondition when the robot will stop balancing
-     * @return true if still balancing, false if finished balancing
-     */
-    boolean balance(double highPower, double lowPower, double moveAngle, double oGain, boolean endCondition);
+    //boolean moveNoIMU(double currentPosition, double targetPosition, double rampDownTargetPosition, double rampUpTargetPosition, double maxPower, double lowPower, double moveAngle, doudouble angle, double speed, boolean condition, double pivotAmount);
 
-    //resets the encoders without stopping the robot.
+    //boolean balance(double highPower, double lowPower, double moveAngle, double oGain, boolean endCondition);
+
+    //resets the encoders.
     void softResetEncoder();
 
-    // Resets encoders and stops the motors
+    // Change the mode of the motors to stop and reset encoders
     void resetEncoders();
 
     //Averages the motor encoders
     double averageEncoders();
 
-    //Stops all the motors on the robot
+    //Gets the distance traveled using encoders
+    double getEncoderDistance();
+
+    //Stops all the motors on the drivetrain
     void stop();
 
-    void setPowerZero();
 }
