+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Teleop/TheWizardTeleop.java
@@ -12,7 +12,7 @@ import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
 import org.firstinspires.ftc.teamcode.GamepadPlus;
 import org.firstinspires.ftc.teamcode.Handiness;
-import org.firstinspires.ftc.teamcode.Subsystems.Drivetrain.OmniDirectionalDrive;
+import org.firstinspires.ftc.teamcode.Subsystems.Drivetrain.MecanumDriveTrain;
 import org.firstinspires.ftc.teamcode.Subsystems.Glyph.DualWheelIntake;
 import org.firstinspires.ftc.teamcode.Subsystems.Glyph.IGlyph;
 import org.firstinspires.ftc.teamcode.Subsystems.Glyph.twoWheelIntake;
@@ -57,7 +57,7 @@ public class TheWizardTeleop extends LinearOpMode {
     IGlyph bottomIntake, topIntake;
     ClawThreePoint relic;
     DualWheelIntake intake;
-    OmniDirectionalDrive drive;
+    MecanumDriveTrain drive;
     TwoPointJewelArm jewel;
 
     List<DcMotor> driveMotors;
@@ -226,7 +226,7 @@ public class TheWizardTeleop extends LinearOpMode {
         waitForStart();
         relic = new ClawThreePoint(relic_extension, relic_arm, relic_tilt, relic_claw);
         intake = new DualWheelIntake(rightWheel1, rightWheel2, leftWheel1, leftWheel2, spin, lift, glyphLimit, telemetry);
-        drive = new OmniDirectionalDrive(driveMotors, telemetry);
+        drive = new MecanumDriveTrain(driveMotors, telemetry);
         jewel = new TwoPointJewelArm(pan, tilt, null, telemetry);
         bottomIntake = new twoWheelIntake(leftWheel1, rightWheel1, INTAKE_POWER, OUTAKE_POWER);
         topIntake = new twoWheelIntake(leftWheel2, rightWheel2, INTAKE_POWER, OUTAKE_POWER);
